I don't know what i'mgoing to merge but I hope it won't break anything

+++ ressources/my_put_float.c
@@ -25,13 +25,13 @@ int my_put_float(long double nb, int precision)
     for (int k = 0; k < precision; k++) {
         marge /= 10;
     }
-    nb += my_put_nbr(nb);
+    nb += my_put_nbr((long)nb);
     my_putchar('.');
     compt++;
     nb += marge;
     for (int index = 0; index < precision; index++) {
-        nb *= 10;
-        nb += my_put_nbr(nb);
+        nb *= (long double)10;
+        nb += my_put_nbr((long)nb);
         compt++;
     }
     return compt;
